feat: refactor form validation and submission utilities into request-utils.js
